+++ lua/mason-config.lua
@@ -11,5 +11,6 @@ mason.setup({
       package_pending = "➜",
       package_uninstalled = "✗"
     }
-  }
+  },
+  PATH = "prepend", -- "skip" seems to cause the spawning error
 })